requirements.txt
streamlit
torch
opencv-python
numpy
pillow